Implementation of USB gamepad controller: 4 analog values (8bit) and 8 buttons.
Had to disable watchdog, otherwise we get watchdog reset. Problem is sending of HID report, if host does not ask for an update (which it doesn't in Linux when no program is using joystick) but we try to send one anyway.

I have to figure out, how to determine if host wants or doesn't want an update report.